// src/common/enums/tarea-codigo.enum.ts
export enum TareaCodigo {
  CREACION_CARPETA = 'Creacion_Carpeta',
  MANIFESTACION_INTERES = 'Manifestacion_Interes',
  ADQUISICION_DERECHO = 'Adquisicion_Derecho_Participar',
  PREPARAR_DOC_JURIDICA = 'Preparar_Doc_Juridica',
  PREPARAR_DOC_TECNICA = 'Preparar_Doc_Tecnica',
  PREPARAR_DOC_FINANCIERA = 'Preparar_Doc_Financiera',
  ESTRUCTURACION_ADMIN = 'Estructuracion_Administracion',
  SOLICITUD_PAGO_POLIZA = 'Solicitud_Pago_Poliza',
  PAGO_POLIZA = 'Pago_Poliza',
  ELABORACION_PROPUESTA = 'Elaboracion_Propuesta_Economica',
  VALIDACION_AREA_TECNICA = 'Validacion_Area_Tecnica',
  ENVIO_PROPUESTA = 'Envio_Propuesta',
}